Restructured the bld_runner module.
