+++ crates/bld_runner/src/platform/mod.rs
@@ -1,4 +1,9 @@
-use crate::context::{Container, Machine};
+mod container;
+mod machine;
+
+pub use container::*;
+pub use machine::*;
+
 use anyhow::Result;
 use bld_core::execution::Execution;
 use std::sync::Arc;
