+++ crates/bld_runner/src/sync/mod.rs
@@ -1,7 +1,3 @@
-mod pipeline;
-mod platform;
-mod runner;
-
-pub use pipeline::*;
-pub use platform::*;
-pub use runner::*;
+pub mod builder;
+pub mod pipeline;
+pub mod runner;
